Lindblad superoperator for Dicke systems
