+++ qutip/models/tests/test_dicke.py
